Fill-in the blank lyrics
Learnt: while True loop.
while True: Used when we need the starting thing - mssg, input etc - to be repeated or in loop. Or used when we need things to be in loop all the time EXCEPT.
Fill in the blank, repeats the question/lyric until we get it right. Only the concept. Not whole lyrics.