Number Guessing Game
As it says. Picks a number between 1 to 100. Asks the user to guess the number. They are given a limited amount of tries. If they suceed they get a well done and are asked if want to play again. If not, they are told the answer and asked to try again? 
Uses same concept as before. while loop. if, elif, else statements. continue, break functions.
Update: From new lessons, I have added an import random function, so you can play the game multiple times.