List Generator
Learnt: Range
Used for loop in range (starting value, ending value, increment).
Prints a list from starting value given by the user to ending value given by user with increment as differences. Increment can be negative or positive but the starting and the ending value should be accordingly as well.