Multiplication Quiz
Used for loop.
Asks user which table they would like to be quizzed on. Asks the table and counts score. If you answer a question right, it says well done and moves on to the next question. If you get it wrong, it says nope and gives you the right answer and then proceeds onto the next question. In the end it gives you score out of 10.