Character Generator
Uses subroutines, while loop and if statements, basic math functions AND COLOR CODE!
Asks you to name your character and then when subroutine is called out picks 2 random numbers uses some operations on them and then gives the value as the character's health score. 
Also, it uses color codes too! If the health is below 2/3rd of the max health possible, it shows health in red color, and if the health is above or equal to 32, it goes green color. Anything between 16 and 32 is color coded yellow. It's random, it's fun.