Sound Player
Learnt: New libraries- os, time, pygame.
Used new libraries, subroutines, loops and functions
As it says. Works like an old ipod. Shows a menu page and asks user to press 1 to play the song, press 2 to exit and press any other button to view the menu again. If user clicks 1, it clears the prev screen, starts playing the audio and gives the message to press 2 to pause playback and go back to the menu. 
It was tough since the vid did not explain the pygame function, plus pygame site gives down error whenever I try to open it. I always run os and other libraries in replit since it doesn't work in pycharm.